pyproject.toml
[project]
name = "android-vendor-crowdin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
]
